requirements.txt
Django
django-markdownx